requirements.txt
transformers
hydra-core
datasets
evaluate
scikit-learn
rank_bm25
accelerate
nltk
spacy
dppy
openai